Update test suite to match current implementation
We’re delegating escaping and sanitisation to `ActionView#sanitize`. The
last commit is 6 years old and the rails dependency is specified as >
3.1 so it’s natural that the test suite is no longer up-to-date.

Current behaviour on Rails 7.0.3.1:

1. Self-closing tags no longer has `/` before `>`.
2. & is escaped as `%amp;`.
3. Save part of malicious content is rendered.
4. `target` attribute is rendered before `href`.
5. `+` is escaped in mailto link.

Notice that 2. makes auto linking emails with & behave differently. This
commit does not change the behaviour but documents the difference in
behaviour. It is possible that this is a regression that should be
addressed.